permission: Rename files to follow the prefix-name rule
Renames the permission files to xdp-permission.
