+++ src/background.c
@@ -31,7 +31,7 @@
 #include "background.h"
 #include "xdp-background-monitor.h"
 #include "flatpak-instance.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 #include "xdp-request.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
