+++ src/gamemode.c
@@ -23,7 +23,7 @@
 #include "config.h"
 
 #include "xdp-call.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
