+++ src/meson.build
@@ -84,7 +84,6 @@ xdg_desktop_portal_sources = files(
   'network-monitor.c',
   'notification.c',
   'open-uri.c',
-  'permissions.c',
   'pipewire.c',
   'power-profile-monitor.c',
   'print.c',
@@ -101,6 +100,7 @@ xdg_desktop_portal_sources = files(
   'xdp-app-launch-context.c',
   'xdp-background-monitor.c',
   'xdp-call.c',
+  'xdp-permissions.c',
   'xdp-portal-impl.c',
   'xdp-request.c',
   'xdp-session.c',
