+++ src/notification.c
@@ -30,7 +30,7 @@
 
 #include "notification.h"
 #include "xdp-call.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 #include "xdp-request.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
