+++ src/realtime.c
@@ -27,7 +27,7 @@
 
 #include "realtime.h"
 #include "xdp-call.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
