+++ src/screen-cast.c
@@ -28,7 +28,7 @@
 #include "screen-cast.h"
 #include "remote-desktop.h"
 #include "xdp-request.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 #include "pipewire.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
