+++ src/wallpaper.c
@@ -29,7 +29,7 @@
 #include <gio/gunixfdlist.h>
 
 #include "wallpaper.h"
-#include "permissions.h"
+#include "xdp-permissions.h"
 #include "xdp-request.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
