+++ src/xdp-permissions.c
@@ -24,7 +24,7 @@
 
 #include <string.h>
 
-#include "permissions.h"
+#include "xdp-permissions.h"
 
 static XdpDbusImplPermissionStore *permission_store = NULL;
